15개
49S,120S,397S,128S,332S,443S,452S,
550S,27D,36D,61D,63D,76D,30X,302S

+++ RC_DPU_PI_200M_4/inc/icd_parser.h
@@ -161,7 +161,7 @@ extern RECV_SETTING DPU_STATUS;
 extern HW_CHECK BIT_STATUS;
 extern HW_CHECK PBIT_STATUS;
 extern RF_SETTING RF_STATUS;
-extern int spec_packet_size;
+extern uint16_t spec_packet_size;
 
 //////// Function Definitions ////////
 ICD_HEADER ParserTCP(uint8_t *recv_buffer, uint16_t packet_len);
